refactor(Hooks): Make `useHideShow` hook generic
